app/requirements.txt
opencv-python-headless
boto3
Flask
Flask-Compress
numpy
pandas==2.1.4
pillow
requests
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.3.2
torchvision==0.16.0
ultralytics==8.2.52
ultralytics-thop==2.0.0
asyncio